remove external files args
makes the API more of a pain to use and is not useful for our use.
subtitles files should probably be in an external file anyway.
audio files should also be checked in another call to dakara_check_avf.
you could argue that a second internal audio track could be the
instrumental track but it's less explicit and therefore error-prone (you
could upload a file with two audio tracks by mistake, in which case the
second track will not be the instrumental version)